Refactor dependency file handling.

+++ tlmgr-install.sh
@@ -3,14 +3,20 @@
 # SPDX-FileCopyrightText: Copyright (c) 2024-2025 Yegor Bugayenko
 # SPDX-License-Identifier: MIT
 
-set -ex
+set -ex -o pipefail
 
 self=$(dirname "$0")
 
 tlmgr option repository ctan
 tlmgr --verify-repo=none update --self
 
-IFS=', ' read -r -a packages <<< "$(cut -d' ' -f2 "${self}/DEPENDS.txt" | uniq)"
+deps="${self}/../DEPENDS.txt"
+if [ ! -e "${deps}" ]; then
+    echo "There is not file with dependencies, why?"
+    echo "Expecting it to be here: ${deps}"
+    exit 1
+fi
+IFS=', ' read -r -a packages <<< "$(cut -d' ' -f2 "${deps}" | uniq)"
 
 tlmgr --verify-repo=none install "${packages[@]}"
 tlmgr --verify-repo=none update "${packages[@]}"
